Primeira Etapa

+++ libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: rosantan <rosantan@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/20 17:10:51 by rosantan          #+#    #+#             */
-/*   Updated: 2022/03/02 19:08:00 by rosantan         ###   ########.fr       */
+/*   Updated: 2022/03/09 19:53:03 by rosantan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*str;
 
 	i = 0;
