+++ libft/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: rosantan <rosantan@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/24 18:55:11 by rosantan          #+#    #+#             */
-/*   Updated: 2022/03/09 11:25:26 by rosantan         ###   ########.fr       */
+/*   Updated: 2022/03/09 19:53:52 by rosantan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,9 +26,9 @@ int	ft_strlen(char *str)
 
 size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
-	int	i;
-	int	sizecpy;
-	int	sizedst;
+	int		i;
+	int		sizecpy;
+	size_t	sizedst;
 
 	i = 0;
 	sizedst = ft_strlen(dst);
